requirements.txt
google-generativeai>=0.7.0
numpy>=1.22.0
opencv-python>=4.7.0
PyAudio>=0.2.13
webrtcvad>=2.0.10 